Break bo_ramp HLA in several modules.
